+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid.stories.js
@@ -120,7 +120,7 @@ const newPerson = () => {
   };
 };
 
-const makeData = (...lens) => {
+export const makeData = (...lens) => {
   const makeDataLevel = (depth = 0) => {
     const len = lens[depth];
     return range(len).map(() => ({
@@ -210,7 +210,7 @@ export const BasicUsage = () => {
     data,
   });
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const InitialLoad = () => {
@@ -237,7 +237,7 @@ export const InitialLoad = () => {
     isFetching,
   });
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const InfiniteScroll = () => {
@@ -270,7 +270,7 @@ export const InfiniteScroll = () => {
 
   return (
     <Wrapper>
-      <Datagrid {...datagridState} />;
+      <Datagrid datagridState={{ ...datagridState }} />;
     </Wrapper>
   );
 };
@@ -286,7 +286,7 @@ export const TenThousandEntries = () => {
     useInfiniteScroll
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 const DatagridPagination = ({ state, setPageSize, gotoPage, rows }) => {
@@ -320,7 +320,7 @@ export const WithPagination = () => {
     DatagridPagination,
   });
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 export const NestedRows = () => {
   const columns = React.useMemo(() => defaultHeader, []);
@@ -333,7 +333,7 @@ export const NestedRows = () => {
     useNestedRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 export const ExpandedRow = () => {
   const expansionRenderer = ({ row }) => <div>Content for {row.id}</div>;
@@ -350,7 +350,7 @@ export const ExpandedRow = () => {
     useExpandedRow
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const NestedTable = () => {
@@ -370,7 +370,7 @@ export const NestedTable = () => {
 
   const expansionRenderer = () => (
     <div className="carbon-nested-table">
-      <Datagrid {...nestedDatagridState} />
+      <Datagrid datagridState={{ ...nestedDatagridState }} />
     </div>
   );
 
@@ -385,7 +385,7 @@ export const NestedTable = () => {
     useExpandedRow
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const ClickableRow = () => {
@@ -400,7 +400,7 @@ export const ClickableRow = () => {
     useOnRowClick
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const IsHoverOnRow = () => {
@@ -431,7 +431,7 @@ export const IsHoverOnRow = () => {
     useRowIsMouseOver
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const SelectableRow = () => {
@@ -445,7 +445,7 @@ export const SelectableRow = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const RadioSelect = () => {
@@ -467,7 +467,7 @@ export const RadioSelect = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const HideSelectAll = () => {
@@ -482,7 +482,7 @@ export const HideSelectAll = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const SortableColumns = () => {
@@ -496,7 +496,7 @@ export const SortableColumns = () => {
     useSortableColumns
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const RightAlignedColumns = () => {
@@ -525,7 +525,7 @@ export const RightAlignedColumns = () => {
     useColumnRightAlign
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 const DatagridActions = (datagridState) => {
@@ -618,7 +618,7 @@ export const DatagridActionsToolbar = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const SelectItemsInAllPages = () => {
@@ -649,7 +649,7 @@ export const SelectItemsInAllPages = () => {
 
   return (
     <>
-      <Datagrid {...datagridState} />
+      <Datagrid datagridState={{ ...datagridState }} />
       <h3>Doc in Notes...</h3>
       <p>{`Are all selected across all pages? - ${areAllSelected}`}</p>
     </>
@@ -682,7 +682,7 @@ export const CustomizingColumns = () => {
 
   return (
     <>
-      <Datagrid {...datagridState} />
+      <Datagrid datagridState={{ ...datagridState }} />
       <div>
         Hidden column ids:
         <pre>{JSON.stringify(datagridState.state.hiddenColumns, null, 2)}</pre>
@@ -742,7 +742,7 @@ export const RowSizeDropdown = () => {
 
   return (
     <Wrapper>
-      <Datagrid {...datagridState} />
+      <Datagrid datagridState={{ ...datagridState }} />
     </Wrapper>
   );
 };
@@ -766,7 +766,7 @@ export const LeftPanel = () => {
 
   return (
     <Wrapper>
-      <Datagrid {...datagridState} />
+      <Datagrid datagridState={{ ...datagridState }} />
     </Wrapper>
   );
 };
@@ -803,7 +803,7 @@ export const BatchActions = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 export const DisableSelectRow = () => {
@@ -822,7 +822,7 @@ export const DisableSelectRow = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
 // export { StickyActionsColumn };
@@ -916,5 +916,5 @@ export const TopAlignment = () => {
     useSelectRows
   );
 
-  return <Datagrid {...datagridState} />;
+  return <Datagrid datagridState={{ ...datagridState }} />;
 };