+++ config/babel-preset-ibm-cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.13](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/babel-preset-ibm-cloud-cognitive@0.14.12...babel-preset-ibm-cloud-cognitive@0.14.13) (2022-04-19)
+
+**Note:** Version bump only for package babel-preset-ibm-cloud-cognitive
+
+
+
+
+
 ## [0.14.12](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/babel-preset-ibm-cloud-cognitive@0.14.11...babel-preset-ibm-cloud-cognitive@0.14.12) (2022-03-29)
 
 **Note:** Version bump only for package babel-preset-ibm-cloud-cognitive